Follow up fixes for debug libraries (#388)
* Bring in libraries on Windows via cpp file

This is easier to manage debug vs. release configs

* Visual Studio is gnu 11 by default

* Remove VS_project.  Get SDK example projects cmake building in Windows

* Fixed example project CMakeLists.txt

* Fix dependency

* Make common project file

* Fix issue with wrong static lib type

Also starting to refactor the sdk compilation cmake file

* Root level cmake working with cltool

Will attempt to get example projects working next

* Almost have example projects working

* Example projects working

* Lib include path was not working

* No need for debug release handling

* Update instructions for Release builds

The first cmake commands creates a makefile which is used for command line builds and thus requires the configuration. The configuration is then removed from the second cmake command.

* Remove debugging code

* Have to use --config

Apparently you have to pass Release or Debug in both steps for cmake

* Rebuild Windows libraries as /mt

Something in the chain was /mt so the static libs had to be rebuilt with that

* Remove SET(BUILD_INERTIAL_SENSE_SDK 1)

* Fix debug and release libs and cmake

CMake requires a very special syntax to set the libs in Debug or Release config

* Allow multiple project files

* Refactor common cmake

* Ignore example projects

* Was using the wrong usb lib

* Got yaml-cpp building in the SDK

No need to bring these files in from elsewhere, and MIT license so easy to add to the project

* Remove not exists path

* Remove libusb

* Remove static libs

Yaml-cpp is bundled as source, libusb is a submodule for Windows and is brought in as a package library on Linux

* Add threads

* Update CMakeCommon.txt

* Example of non-common cmake pattern

* WIP

* Comment out debug code

* Remove files

* Remove submodule libusb

* Remove submodule libusb

* Re-adding submodule libusb

* Linux libusb config.h file

New remote for libusb pointing to an InertialSense repository that can be maintained in the future

* Build Linux libusb

* Link winsock on Windows for loginspector python

* Put udev back

* Ignore non windows compiler directive

* Try moving target libraries before pybind call

* Simplify common cmake setup

* Revert to dll linkage

* Remove unused CMakeList.txt from LogInspector.

* Fix reference to navpp.exe in supernpp.py

* Fix navpp.exe call from supernpp.py

* Remove duplicate include

* Revert test

* Try to fix ros build

* Remove stray comma

* Revert from commit at 12pm today

* Fix compiler warning

* Remove dead code

---------

Co-authored-by: waltjohnson <walt@inertialsense.com>